+++ agentMET4FOF_ml_extension/datastream.py
@@ -1,8 +1,8 @@
-from ..agents import AgentMET4FOF
+from agentMET4FOF.agents import AgentMET4FOF
 from sklearn.model_selection import KFold
 
 
-class DataStreamAgent(AgentMET4FOF):
+class ML_DataStreamAgent(AgentMET4FOF):
     def init_parameters(self, data_name="unnamed_data", train_mode={"Prequential","Kfold5","Kfold10"}, x=None, y=None):
         #if data_name is provided, will assign to pre-made datasets,
         #otherwise for custom dataset, three parameters : data_name, x and y will need to be assigned
