+++ agentMET4FOF_ml_extension/tutorials/with_bnn/ml_experiment_simple.py
@@ -10,29 +10,19 @@ consisting of pipelines and subsets called chains, can be viewed and compared.
 """
 
 
-from agentMET4FOF.agents import AgentMET4FOF, AgentNetwork, MonitorAgent, AgentPipeline
-from agentMET4FOF.develop.datastream import *
-from agentMET4FOF.develop.evaluator import *
-from agentMET4FOF.develop.ML_Experiment import *
+from agentMET4FOF.agents import AgentNetwork, MonitorAgent
+from agentMET4FOF_ml_extension.datastream import *
+from agentMET4FOF_ml_extension.evaluator import *
+from agentMET4FOF_ml_extension.ML_Experiment import *
+from agentMET4FOF_ml_extension.ml_uncertainty.bnn import BNN_Model
+from agentMET4FOF_ml_extension.ml_uncertainty.evaluate_pred_unc import *
+from agentMET4FOF_ml_extension.agents import AgentPipeline
 
-from sklearn.model_selection import KFold
 from sklearn import datasets
-from sklearn.metrics import f1_score
-
-from sklearn.neighbors import KNeighborsClassifier as KNN
-from sklearn.decomposition import PCA
-from sklearn.preprocessing import StandardScaler, RobustScaler, MinMaxScaler, MaxAbsScaler,PowerTransformer
-from sklearn.svm import SVC
-from sklearn.linear_model import LogisticRegression
-from sklearn.model_selection import ParameterGrid
-from pprint import pprint
-import copy
-
-from agentMET4FOF.ml_uncertainty.bnn import BNN_Model
-from agentMET4FOF.ml_uncertainty.evaluate_pred_unc import *
+from sklearn.preprocessing import StandardScaler
 
 def main():
-    agentNetwork = AgentNetwork()
+    agentNetwork = AgentNetwork(dashboard_extensions=Dashboard_ML_Experiment)
 
     ml_exp_name = "simple"
 
@@ -44,15 +34,14 @@ def main():
 
 
     #init
-    datastream_agent = agentNetwork.add_agent(agentType=DataStreamAgent)
-    evaluation_agent = agentNetwork.add_agent(agentType=EvaluationAgent)
+    datastream_agent = agentNetwork.add_agent(agentType=ML_DataStreamAgent)
+    evaluation_agent = agentNetwork.add_agent(agentType=ML_EvaluatorAgent)
 
     datastream_agent.init_parameters(data_name="iris", x=datasets.load_iris().data,y=datasets.load_iris().target)
     evaluation_agent.init_parameters([f1_score,p_acc_unc,avg_unc],[{"average":'micro'},{},{}], ML_exp=True)
 
     #setup ml experiment
-    ml_experiment = ML_Experiment(datasets=[datastream_agent], pipelines=[ML_Agent_pipelines_A], evaluation=[evaluation_agent], name=ml_exp_name, train_mode="Kfold5")
-    agentNetwork.get_agent("Logger").set_ml_experiment(ml_experiment)
+    ml_experiment = ML_Experiment(agentNetwork, datasets=[datastream_agent], pipelines=[ML_Agent_pipelines_A], evaluation=[evaluation_agent], name=ml_exp_name, train_mode="Kfold5")
 
     #optional: connect evaluation agent to monitor agent
     monitor_agent = agentNetwork.add_agent(agentType=MonitorAgent)
